backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
redis>=5.2.0
aiomysql>=0.2.0
openai>=1.0.0
httpx>=0.27.0
